refactor(core-android): move contact module to android module
