+++ legacy/ui/legacy/build.gradle.kts
@@ -13,7 +13,7 @@ dependencies {
     implementation(projects.feature.mail.account.api)
     implementation(projects.mail.common)
     implementation(projects.uiUtils.toolbarBottomSheet)
-    implementation(projects.core.contact)
+    implementation(projects.core.android.contact)
 
     implementation(projects.core.featureflags)
     implementation(projects.core.ui.theme.api)
