+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/MessageCompose.java
@@ -118,7 +118,7 @@ import com.fsck.k9.ui.helper.SizeFormatter;
 import com.fsck.k9.ui.messagelist.DefaultFolderProvider;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.textview.MaterialTextView;
-import net.thunderbird.core.contact.ContactIntentHelper;
+import net.thunderbird.core.android.contact.ContactIntentHelper;
 import net.thunderbird.core.ui.theme.manager.ThemeManager;
 import net.thunderbird.feature.search.LocalSearch;
 import org.openintents.openpgp.OpenPgpApiManager;
